Added early exit and command permission guarding
**Description:**

Added early exit for when interaction is not a command.

Added permission guarding of interaction commands by implementing the recently created CaseModel to check if the user is has permission.

If the user is in blacklist, or they are not whitelisted, an ephemeral message will be sent to them with the error.

At this time of the commit, the library "discord" - discord.js but for typescript - complains in the event if the return type is not "Awaitable<void>". Whilst the code in this commit ensures "Awaitable<void>" is returned, a query has been sent over to the "discordx" support server asking about this.